[Improve][Jdbc] Improve catalog connection cache
